backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyjwt==2.8.1
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
gunicorn==21.2.0
mlflow==2.10.0
joblib==1.3.2
